Funcionando pero sin crear base de datos

+++ app/src/main/java/com/example/alkewalletfinal/model/response/UserResponse.kt
@@ -3,15 +3,15 @@ package com.example.alkewalletfinal.model.response
 import com.google.gson.annotations.SerializedName
 
 data class UserResponse(
-    val id: Int,
+    val id: Long,
     @SerializedName("first_name")
     val firstName: String,
     @SerializedName("last_name")
     val lastName: String,
     val email: String,
     val password: String,
-    val points: Long,
-    val roleid: Long,
+    val points: Int,
+    val roleid: Int,
     val createdAt: String,
     val updateAt: String
 )
